update api sync

+++ packages/api-database/source/models/state.ts
@@ -16,7 +16,7 @@ export class State {
 	public height!: string;
 
 	@Column({
-		type: "bigint",
+		type: "numeric",
 	})
 	public supply!: string;
 }
